Abstract heatmap track labeling to new module
