+++ src/js/annotations/heatmap-lib.js
@@ -1,0 +1,35 @@
+
+var reservedTrackKeys = [
+  'name', 'start', 'length', 'trackIndex', 'trackIndexOriginal', 'color'
+];
+
+/**
+ * Get label text for displayed tracks from annotation container metadata,
+ * heatmap keys, or annotation container keys
+ */
+function getLabels(ideo) {
+  var annotKeys, labels, heatmaps, i;
+
+  if (ideo.rawAnnots.metadata) {
+    labels = ideo.rawAnnots.metadata.trackLabels;
+  } else if (ideo.config.heatmaps) {
+    labels = [];
+    heatmaps = ideo.config.heatmaps;
+    for (i = 0; i < heatmaps.length; i++) {
+      labels.push(heatmaps[i].key);
+    }
+  } else {
+    annotKeys = ideo.rawAnnots.keys.slice(0);
+    labels = annotKeys.filter(d => !reservedTrackKeys.includes(d));
+  }
+
+  if (ideo.displayedTrackIndexes) {
+    labels = labels.filter(function(d, i) {
+      return ideo.displayedTrackIndexes.includes(i + 1);
+    });
+  }
+
+  return labels;
+}
+
+export default getLabels;
