+++ src/js/annotations/track-labels.js
@@ -6,11 +6,9 @@
 
 import * as d3selection from 'd3-selection';
 
-var d3 = Object.assign({}, d3selection);
+import getLabels from './heatmap-lib';
 
-var reservedTrackKeys = [
-  'name', 'start', 'length', 'trackIndex', 'trackIndexOriginal', 'color'
-];
+var d3 = Object.assign({}, d3selection);
 
 /**
  * Start a timer that, upon expiring, hides the track label.
@@ -46,37 +44,6 @@ function writeTrackLabelContainer(ideo) {
     .style('background', 'white')
     .style('line-height', '10px')
     .style('z-index', '9000')
-}
-
-/**
- * Get label text for displayed tracks from annotation container metadata,
- * heatmap keys, or annotation container keys
- */
-function getLabels(ideo) {
-  var annotKeys, labels, heatmaps, i;
-
-  if (ideo.rawAnnots.metadata) {
-    labels = ideo.rawAnnots.metadata.trackLabels;
-  } else if (ideo.config.heatmaps) {
-    labels = [];
-    heatmaps = ideo.config.heatmaps;
-    for (i = 0; i < heatmaps.length; i++) {
-      labels.push(heatmaps[i].key);
-    }
-  } else {
-    annotKeys = ideo.rawAnnots.keys.slice(0);
-    labels = annotKeys.filter(d => !reservedTrackKeys.includes(d));
-  }
-
-  if (ideo.displayedTrackIndexes) {
-    labels = labels.filter(function(d, i) {
-      return ideo.displayedTrackIndexes.includes(i + 1);
-    });
-  }
-
-  labels = labels.join('<br>');
-
-  return labels;
 }
 
 /**
@@ -131,6 +98,7 @@ function showTrackLabel(trackCanvas, ideo) {
     clearTimeout(ideo.hideTrackLabelTimeout);
 
     labels = getLabels(ideo);
+    labels = labels.join('<br>');
 
     // Clear any previous positioning, write track label text to DOM
     d3.select(ideo.config.container + ' #_ideogramTrackLabel')
